dev merge

+++ modules/ui_common.py
@@ -68,15 +68,22 @@ def delete_files(js_data, files, all_files, index):
         start_index = index
     deleted = []
     all_files = [f.split('/file=')[1] if 'file=' in f else f for f in all_files] if isinstance(all_files, list) else []
+    all_files = [os.path.normpath(f) for f in all_files]
+    reference_dir = os.path.join('models', 'Reference')
     for _image_index, filedata in enumerate(files, start_index):
         try:
-            fn = filedata['name']
+            fn = os.path.normpath(filedata['name'])
+            if reference_dir in fn:
+                shared.log.warning(f'Delete: file="{fn}" not allowed')
+                continue
             if os.path.exists(fn) and os.path.isfile(fn):
                 deleted.append(fn)
                 os.remove(fn)
                 if fn in all_files:
                     all_files.remove(fn)
-                shared.log.info(f'Delete: image="{fn}"')
+                    shared.log.info(f'Delete: image="{fn}"')
+                else:
+                    shared.log.warning(f'Delete: image="{fn}" ui mismatch')
             base, _ext = os.path.splitext(fn)
             desc = f'{base}.txt'
             if os.path.exists(desc) and os.path.isfile(desc):
@@ -333,7 +340,7 @@ def create_refresh_button(refresh_component, refresh_method, refreshed_args = No
         return gr.update(**args)
 
     refresh_button = ui_components.ToolButton(value=ui_symbols.refresh, elem_id=elem_id, visible=visible)
-    refresh_button.click(fn=refresh, inputs=[], outputs=[refresh_component])
+    refresh_button.click(fn=refresh, inputs=[], outputs=[refresh_component], show_progress=False)
     return refresh_button
 
 
@@ -344,7 +351,26 @@ def create_override_inputs(tab): # pylint: disable=unused-argument
     return override_settings
 
 
-def connect_reuse_seed(seed: gr.Number, reuse_seed: gr.Button, generation_info: gr.Textbox, is_subseed, subseed_strength=None):
+def reuse_seed(seed_component: gr.Number, reuse_button: gr.Button, subseed:bool=False):
+    def reuse_click(selected_gallery_index):
+        selected_gallery_index = int(selected_gallery_index)
+        from modules import processing
+        if processing.processed is None:
+            seed = -1
+        elif selected_gallery_index >= len(processing.processed.all_seeds):
+            selected_gallery_index -= len(processing.processed.images) - len(processing.processed.all_seeds) # if we have more images than seeds it is likely the grid image
+            seed = processing.processed.all_seeds[selected_gallery_index] if not subseed else processing.processed.all_subseeds[selected_gallery_index]
+        elif len(processing.processed.all_seeds) > 0:
+            seed = processing.processed.all_seeds[0] if not subseed else processing.processed.all_subseeds[0]
+        else:
+            seed = -1
+        shared.log.debug(f'Reuse seed: index={selected_gallery_index} seed={seed} subseed={subseed}')
+        return seed
+
+    reuse_button.click(fn=reuse_click, _js="selected_gallery_index", inputs=[seed_component], outputs=[seed_component], show_progress=False)
+
+
+def connect_reuse_seed(seed: gr.Number, reuse_seed_btn: gr.Button, generation_info: gr.Textbox, is_subseed, subseed_strength=None):
     """ Connects a 'reuse (sub)seed' button's click event so that it copies last used
         (sub)seed value from generation info the to the seed field. If copying subseed and subseed strength
         was 0, i.e. no variation seed was used, it copies the normal seed value instead."""
@@ -372,9 +398,9 @@ def connect_reuse_seed(seed: gr.Number, reuse_seed: gr.Button, generation_info: 
             return [restore_seed, gr_show(False)]
     dummy_component = gr.Number(visible=False, value=0)
     if subseed_strength is None:
-        reuse_seed.click(fn=copy_seed, _js="(x, y) => [x, selected_gallery_index()]", show_progress=False, inputs=[generation_info, dummy_component], outputs=[seed, dummy_component])
+        reuse_seed_btn.click(fn=copy_seed, _js="(x, y) => [x, selected_gallery_index()]", show_progress=False, inputs=[generation_info, dummy_component], outputs=[seed, dummy_component])
     else:
-        reuse_seed.click(fn=copy_seed, _js="(x, y) => [x, selected_gallery_index()]", show_progress=False, inputs=[generation_info, dummy_component], outputs=[seed, dummy_component, subseed_strength])
+        reuse_seed_btn.click(fn=copy_seed, _js="(x, y) => [x, selected_gallery_index()]", show_progress=False, inputs=[generation_info, dummy_component], outputs=[seed, dummy_component, subseed_strength])
 
 
 def update_token_counter(text):
